requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
mysqlclient==2.1.1
PyJWT==2.6.0
pyrsistent==0.19.3
PySocks==1.7.1
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
redis==4.5.4
requests==2.28.2
sqlparse==0.4.3
stripe==5.4.0
uritemplate==4.1.1
urllib3==1.26.15
